-Move ValidatorConfig from wrong package to correct package -Implement light-rest-4j #64 for Swagger ValidatorHandler
